git-svn-id: http://svn.code.sf.net/p/winmerge/code/branches/Perry@3697 a4945e5c-c7c5-4958-bd9a-f56ba18d93a3

+++ Src/LocationView.cpp
@@ -191,8 +191,8 @@ void CLocationView::OnDraw(CDC* pDC)
 
 	// Draw bar outlines
 	CPen* oldObj = (CPen*)pDC->SelectStockObject(BLACK_PEN);
-	pDC->Rectangle(m_nLeftBarLeft, Y_OFFSET - 1, m_nLeftBarRight, LineInPix * nbLines + Y_OFFSET);
-	pDC->Rectangle(m_nRightBarLeft, Y_OFFSET - 1, m_nRightBarRight, LineInPix * nbLines + Y_OFFSET);
+	pDC->Rectangle(m_nLeftBarLeft, Y_OFFSET - 1, m_nLeftBarRight, (int)(LineInPix * nbLines) + Y_OFFSET);
+	pDC->Rectangle(m_nRightBarLeft, Y_OFFSET - 1, m_nRightBarRight, (int)(LineInPix * nbLines) + Y_OFFSET);
 	pDC->SelectObject(oldObj);
 
 	while (true)
